[Modify] 단순 주석 및 타입 위치 수정
