--- src/components/common/atoms/Select.d.ts
@@ -1,8 +1,0 @@
-import { IUseSelectResult } from "../../../hooks/useSelect.d";
-
-/** \# Select 타입 */
-export interface ISelect {
-  width: string;
-  list: string[];
-  select: IUseSelectResult;
-}
