+++ src/styles/GlobalStyle.tsx
@@ -12,11 +12,10 @@ const GlobalStyle = createGlobalStyle`
     font-size: 62.5%; //> 1rem = 10px;  
   }
 
-  //^IE
+  //# SELECT 기본 스타일 제거
   select::-ms-expand {
     display: none;
   }
-  //# SELECT 기본 스타일 제거
   select {
     -webkit-appearance: none;
     -moz-appearance: none;
